rename & update wolfram solver
